Ensure per_epoch processing always runs.
Previously, it was running _after_ a state transition, not before it
with the slot processing.
